Explicitely cast NSIntegers to avoid warning and problems
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@37916 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormClassInspector.m
@@ -658,9 +658,9 @@ objectValueForTableColumn: (NSTableColumn *)tc
       actionsCount = [[classManager allActionsForClassNamed: currentClass] count];
       
       item = [tabView tabViewItemAtIndex: 1]; // actions;
-      [item setLabel: [NSString stringWithFormat: @"Actions (%ld)",actionsCount]];
+      [item setLabel: [NSString stringWithFormat: @"Actions (%ld)",(long)actionsCount]];
       item = [tabView tabViewItemAtIndex: 0]; // outlets;
-      [item setLabel: [NSString stringWithFormat: @"Outlets (%ld)",outletsCount]];
+      [item setLabel: [NSString stringWithFormat: @"Outlets (%ld)",(long)outletsCount]];
       [tabView setNeedsDisplay: YES];
       
       [self _refreshView];
